Разделил код в main.js на отдельные модули; импортировал нужные модули в точку входа; точку входа подключил в index.html
